requirements.txt
fastapi[standard]==0.115.*
python-dateutil==2.9.*
pydantic-settings==2.8.*
passlib==1.7.4
asyncpg==0.30.0
alembic==1.14.*
SQLAlchemy==2.0.*
sqlalchemy-utils==0.41.*
psycopg2-binary==2.9.*
polars==1.24.*
pandas==2.2.*
pyarrow==19.*
pyjwt==2.10.*
passlib[bcrypt]==1.7.*
